requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.2.3
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.56
langchain-mistralai==0.2.10
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langfuse==2.60.3
langgraph==0.3.34
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.37
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.36.0
numpy==2.2.5
ollama==0.4.8
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pipx==1.5.0
propcache==0.3.1
protobuf==5.29.4
python-dotenv==1.1.0
pytz==2025.2
pywin32==307
PyYAML==6.0.2
pyarrow==19.0.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
requests==2.32.3
requests-toolbelt==1.0.0
referencing==0.36.2
rpds-py==0.24.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
streamlit==1.44.1
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
userpath==1.9.2
uvicorn==0.34.2
watchdog==6.0.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
rich==14.0.0
pydantic
chromadb
sentence-transformers
thefuzz
langchain_experimental
tabulate